# contracts/openapi/music.yaml
openapi: 3.1.0
...
info:
  title: music-view API
  version: 1.0.0
  description: |
    Narzędzie DJ-a do zarządzania biblioteką muzyczną. Rozdział danych:
    `track_catalog` trzyma fakty o utworze, `library_entry` prywatne dane DJ-a.
servers:
  - url: /
    description: aplikacja za proxy /api
...
paths:
...
  /api/catalog/tracks/{spotifyId}:
    get:
      tags: [catalog]
      operationId: getTrack
      summary: Pojedynczy utwór z katalogu
      parameters: [{ $ref: '#/components/parameters/SpotifyId' }]
      responses:
        '200':
          description: Utwór
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrackResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/library/tracks:
    get:
      tags: [library]
      operationId: listLibraryTracks
      summary: Strona wpisów bibliotecznych
      parameters:
        - { name: page, in: query, schema: { type: integer, format: int32, default: 0 } }
        - { name: size, in: query, schema: { type: integer, format: int32, default: 20 } }
      responses:
        '200':
          description: Strona wpisów
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageLibraryEntryResponse' }
    post:
      tags: [library]
      operationId: addLibraryTrack
      summary: Ręczne dodanie utworu do biblioteki
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddLibraryTrackRequest' }
      responses:
        '201':
          description: Wpis utworzony
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LibraryEntryResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/library/tracks/{spotifyId}:
    get:
      tags: [library]
      operationId: getLibraryTrack
      parameters: [{ $ref: '#/components/parameters/SpotifyId' }]
      responses:
        '200':
          description: Wpis biblioteczny
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LibraryEntryResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [library]
      operationId: updateLibraryTrack
      summary: Zmiana danych prywatnych DJ-a
      description: |
        Pole pominięte zostaje bez zmian; pusta wartość czyści pole.
        `version` jest wymagana — nieświeża kończy się 409.
      parameters: [{ $ref: '#/components/parameters/SpotifyId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateLibraryEntryRequest' }
      responses:
        '200':
          description: Wpis po zmianie
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LibraryEntryResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [library]
      operationId: deleteLibraryTrack
      summary: Usunięcie wpisu; katalog zostaje bez zmian
      parameters: [{ $ref: '#/components/parameters/SpotifyId' }]
      responses:
        '204': { description: Usunięto }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/ingest/file:
    post:
      tags: [ingest]
      operationId: ingestFile
      summary: Import utworów z pliku CSV
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200':
          description: Raport importu
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestFileResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /api/ingest/playlist:
    post:
      tags: [ingest]
      operationId: ingestPlaylist
      summary: Import playlisty ze Spotify spod linku
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IngestPlaylistRequest' }
      responses:
        '200':
          description: Raport importu
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestPlaylistResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '502': { $ref: '#/components/responses/ExternalServiceError' }
  /api/ingest/my-playlists:
    post:
      tags: [ingest]
      operationId: ingestMyPlaylists
      summary: Import wszystkich playlist połączonego konta
      description: Awaria pojedynczej playlisty nie przerywa przebiegu.
      responses:
        '200':
          description: Raport zbiorczy
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestMyPlaylistsResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  parameters:
    SpotifyId:
      name: spotifyId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Błąd walidacji
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Zasób nie istnieje
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Konflikt — zasób zmieniony w innym miejscu albo już istnieje
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ExternalServiceError:
      description: Zewnętrzne API odmówiło odpowiedzi
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ErrorResponse:
      type: object
      required: [errorCode, message, timestamp, traceId]
      properties:
        errorCode: { type: string, examples: [TRACK_NOT_FOUND] }
        message: { type: string }
        timestamp: { type: string, format: date-time }
        traceId: { type: [string, 'null'] }
    TrackResponse:
      type: object
      required: [spotifyId, title, artist, album, year, durationMs, popularity, explicit, albumImageUrl,
                 isrc, genreFamily, style, bpm, bpmSource, danceability, musicalKey,
                 camelot, tempoClass, energy, lyricsTheme, descriptionPl, confidence,
                 enrichedAt, modelUsed, enrichVersion]
      properties:
        spotifyId: { type: string }
        title: { type: [string, 'null'] }
        artist: { type: [string, 'null'] }
        album: { type: [string, 'null'] }
        year: { type: [integer, 'null'], format: int32 }
        durationMs: { type: [integer, 'null'], format: int32 }
        popularity: { type: [integer, 'null'], format: int32 }
        explicit: { type: [boolean, 'null'] }
        albumImageUrl: { type: [string, 'null'] }
        isrc: { type: [string, 'null'] }
        genreFamily: { type: [string, 'null'] }
        style: { type: [string, 'null'] }
        bpm: { type: [integer, 'null'], format: int32 }
        bpmSource: { type: [string, 'null'] }
        danceability: { type: [number, 'null'] }
        musicalKey: { type: [string, 'null'] }
        camelot:
          type: [string, 'null']
          description: Liczona z `musicalKey` przez backend — nie kolumna w bazie
        tempoClass: { type: [string, 'null'] }
        energy: { type: [string, 'null'] }
        lyricsTheme: { type: [string, 'null'] }
        descriptionPl: { type: [string, 'null'] }
        confidence: { type: [string, 'null'] }
        enrichedAt: { type: [string, 'null'], format: date-time }
        modelUsed: { type: [string, 'null'] }
        enrichVersion: { type: [integer, 'null'], format: int32 }
...
    PageLibraryEntryResponse:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/LibraryEntryResponse' }
        page: { type: integer, format: int32 }
        size: { type: integer, format: int32 }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer, format: int32 }
...
    LibraryEntryResponse:
      type: object
      required: [id, spotifyId, source, addedAt, djNotes, customTags, rating, djSlotOverride, version,
                 track]
      properties:
        id: { type: integer, format: int64 }
        spotifyId: { type: string }
        source: { type: string }
        addedAt: { type: string, format: date-time }
        djNotes: { type: [string, 'null'] }
        customTags: { type: [array, 'null'], items: { type: string } }
        rating: { type: [integer, 'null'], format: int32 }
        djSlotOverride: { type: [string, 'null'] }
        version: { type: integer, format: int32 }
        track: { $ref: '#/components/schemas/TrackResponse' }
    AddLibraryTrackRequest:
      type: object
      required: [spotifyId, title, artist]
      properties:
        spotifyId: { type: string, pattern: '[A-Za-z0-9_-]{1,64}' }
        title: { type: string, minLength: 1 }
        artist: { type: string, minLength: 1 }
        album: { type: [string, 'null'] }
    UpdateLibraryEntryRequest:
      type: object
      required: [version]
      description: Pole pominięte zostaje bez zmian; pusta wartość czyści pole.
      properties:
        djNotes: { type: [string, 'null'] }
        customTags: { type: [array, 'null'], items: { type: string } }
        rating:
          type: [integer, 'null']
          format: int32
          description: 0 czyści ocenę; poza tym dozwolone 1–5
        djSlotOverride: { type: [string, 'null'] }
        version: { type: integer, format: int32 }
...
    RowErrorResponse:
      type: object
      required: [line, reason]
      properties:
        line: { type: integer, format: int64 }
        reason: { type: string }
    IngestFileResponse:
      type: object
      required: [imported, alreadyExisted, failed]
      properties:
        imported: { type: integer, format: int32 }
        alreadyExisted: { type: integer, format: int32 }
        failed:
          type: array
          items: { $ref: '#/components/schemas/RowErrorResponse' }
...
    IngestPlaylistRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          minLength: 1
          maxLength: 500
          description: Link, URI `spotify:playlist:…` albo samo id playlisty
    SkippedItemResponse:
      type: object
      required: [position, reason]
      properties:
        position: { type: integer, format: int32 }
        reason: { type: string }
    IngestPlaylistResponse:
      type: object
      required: [playlistId, spotifyPlaylistId, name, tracks, imported, alreadyExisted, skipped]
      properties:
        playlistId: { type: [integer, 'null'], format: int64 }
        spotifyPlaylistId: { type: string }
        name: { type: string }
        tracks: { type: integer, format: int32 }
        imported: { type: integer, format: int32 }
        alreadyExisted: { type: integer, format: int32 }
        skipped:
          type: array
          items: { $ref: '#/components/schemas/SkippedItemResponse' }
    FailedPlaylistResponse:
      type: object
      required: [spotifyPlaylistId, name, errorCode, reason]
      properties:
        spotifyPlaylistId: { type: string }
        name: { type: [string, 'null'] }
        errorCode: { type: string }
        reason: { type: string }
    IngestMyPlaylistsResponse:
      type: object
      required: [imported, failed]
      properties:
        imported:
          type: array
          items: { $ref: '#/components/schemas/IngestPlaylistResponse' }
        failed:
          type: array
          items: { $ref: '#/components/schemas/FailedPlaylistResponse' }
...